refactor debug adapter to start separate adapters for Metro and JS debugging
