+++ packages/vscode-extension/src/debugging/DebugAdapter.ts
@@ -1,29 +1,7 @@
-import fs from "fs";
-import path from "path";
-import os from "os";
-import { DebugConfiguration } from "vscode";
-import {
-  DebugSession,
-  InitializedEvent,
-  StoppedEvent,
-  Event,
-  ContinuedEvent,
-  OutputEvent,
-  Thread,
-  TerminatedEvent,
-  ThreadEvent,
-  Source,
-  StackFrame,
-} from "@vscode/debugadapter";
+import { debug, DebugConsoleMode, DebugSession } from "vscode";
+import { DebugSession as DebugAdapterSession, OutputEvent, Source } from "@vscode/debugadapter";
 import { DebugProtocol } from "@vscode/debugprotocol";
 import { Logger } from "../Logger";
-import {
-  inferDAPScopePresentationHintFromCDPType,
-  inferDAPVariableValueForCDPRemoteObject,
-  CDPDebuggerScope,
-} from "./cdp";
-import { CDPSession, CDPSessionDelegate } from "./CDPSession";
-import getArraySlots from "./templates/getArraySlots";
 import { DebugSource } from "./DebugSession";
 
 export type CDPConfiguration = {
@@ -33,7 +11,7 @@ export type CDPConfiguration = {
   breakpointsAreRemovedOnContextCleared: boolean;
 };
 
-const ERROR_RESPONSE_FAIL_TO_RETRIEVE_VARIABLE_ID = 4000;
+const JS_DEBUGGER_TYPE = "com.swmansion.js-debugger";
 
 export function typeToCategory(type: string) {
   switch (type) {
@@ -45,87 +23,39 @@ export function typeToCategory(type: string) {
   }
 }
 
-export class DebugAdapter extends DebugSession implements CDPSessionDelegate {
-  private cdpSession: CDPSession | undefined;
+export class DebugAdapter extends DebugAdapterSession {
+  private cdpDebugSession: DebugSession | null = null;
 
-  private threads: Array<Thread> = [];
-
-  private pausedStackFrames: StackFrame[] = [];
-  private pausedScopeChains: CDPDebuggerScope[][] = [];
-
-  constructor(configuration: DebugConfiguration) {
+  constructor(private vscDebugSession: DebugSession) {
     super();
   }
 
   private connectJSDebugger(cdpConfiguration: CDPConfiguration) {
-    this.cdpSession = new CDPSession(
-      this,
-      cdpConfiguration.websocketAddress,
+    const unsub = debug.onDidStartDebugSession((session) => {
+      if (session.type === JS_DEBUGGER_TYPE) {
+        this.cdpDebugSession = session;
+        unsub.dispose();
+      }
+    });
+    debug.startDebugging(
+      undefined,
       {
-        expoPreludeLineCount: cdpConfiguration.expoPreludeLineCount,
-        sourceMapAliases: cdpConfiguration.sourceMapAliases,
+        type: JS_DEBUGGER_TYPE,
+        name: "React Native JS Debugger",
+        request: "attach",
+        ...cdpConfiguration,
       },
       {
-        breakpointsAreRemovedOnContextCleared:
-          cdpConfiguration.breakpointsAreRemovedOnContextCleared,
+        parentSession: this.vscDebugSession,
+        suppressDebugStatusbar: true,
+        suppressDebugView: true,
+        suppressDebugToolbar: true,
+        suppressSaveBeforeStart: true,
+        consoleMode: DebugConsoleMode.MergeWithParent,
+        compact: true,
       }
     );
   }
-  //#region CDPDelegate
-
-  public onExecutionContextCreated = (threadId: number, threadName: string) => {
-    this.sendEvent(new ThreadEvent("started", threadId));
-    this.threads.push(new Thread(threadId, threadName));
-  };
-
-  public onConnectionClosed = () => {
-    this.sendEvent(new TerminatedEvent());
-  };
-
-  public onDebugSessionReady = () => {
-    this.sendEvent(new InitializedEvent());
-  };
-
-  public onDebuggerPaused = (message: any) => {
-    // this.handleDebuggerPaused(message);
-  };
-
-  public onDebuggerResumed = () => {
-    this.sendEvent(new ContinuedEvent(this.threads[0].id));
-  };
-
-  public onExecutionContextsCleared = () => {
-    const allThreads = this.threads;
-    this.threads = [];
-    // send events for all threads that exited
-    allThreads.forEach((thread) => {
-      this.sendEvent(new ThreadEvent("exited", thread.id));
-    });
-
-    // send event to clear console
-    this.sendEvent(new OutputEvent("\x1b[2J", "console"));
-  };
-
-  public sendOutputEvent = (output: OutputEvent) => {
-    this.sendEvent(output);
-    this.sendEvent(new Event("RNIDE_consoleLog", { category: output.body.category }));
-  };
-
-  public sendStoppedEvent = (
-    pausedStackFrames: StackFrame[],
-    pausedScopeChains: CDPDebuggerScope[][],
-    reason: string,
-    exceptionText?: string,
-    isFatal?: string
-  ) => {
-    this.pausedStackFrames = pausedStackFrames;
-    this.pausedScopeChains = pausedScopeChains;
-
-    this.sendEvent(new StoppedEvent(reason, this.threads[0].id, exceptionText));
-    this.sendEvent(new Event("RNIDE_paused", { reason, isFatal }));
-  };
-
-  //#endregion
 
   logCustomMessage(message: string, category: string, source?: DebugSource) {
     const output = new OutputEvent(message, typeToCategory(category));
@@ -141,256 +71,22 @@ export class DebugAdapter extends DebugSession implements CDPSessionDelegate {
     this.sendEvent(output);
   }
 
-  //#region DAP Implementation
-
-  protected initializeRequest(
-    response: DebugProtocol.InitializeResponse,
-    args: DebugProtocol.InitializeRequestArguments
-  ): void {
-    response.body = response.body || {};
-
-    this.sendResponse(response);
-  }
-
-  protected launchRequest(
-    response: DebugProtocol.LaunchResponse,
-    args: DebugProtocol.LaunchRequestArguments
-  ): void {
-    // Implement launching the debugger
-    this.sendResponse(response);
-  }
-
-  protected async setBreakPointsRequest(
-    response: DebugProtocol.SetBreakpointsResponse,
-    args: DebugProtocol.SetBreakpointsArguments
-  ): Promise<void> {
-    if (!this.cdpSession) {
-      Logger.warn("[DebugAdapter] [setBreakPointsRequest] The CDPSession was not initialized yet");
-      this.sendResponse(response);
-      return;
-    }
-
-    const sourcePath = args.source.path;
-    if (!sourcePath) {
-      this.sendResponse(response);
-      return;
-    }
-
-    const resolvedBreakpoints = await this.cdpSession.handleSetBreakpointRequest(
-      sourcePath,
-      args.breakpoints
-    );
-
-    // send back the actual breakpoint positions
-    response.body = {
-      breakpoints: resolvedBreakpoints,
-    };
-    this.sendResponse(response);
-  }
-
-  protected threadsRequest(response: DebugProtocol.ThreadsResponse): void {
-    response.body = {
-      threads: this.threads,
-    };
-    this.sendResponse(response);
-  }
-
-  protected stackTraceRequest(
-    response: DebugProtocol.StackTraceResponse,
-    args: DebugProtocol.StackTraceArguments
-  ): void {
-    response.body = response.body || {};
-    response.body.stackFrames = this.pausedStackFrames;
-    this.sendResponse(response);
-  }
-
-  protected async scopesRequest(
-    response: DebugProtocol.ScopesResponse,
-    args: DebugProtocol.ScopesArguments
-  ): Promise<void> {
-    if (!this.cdpSession) {
-      Logger.warn("[DebugAdapter] [scopesRequest] The CDPSession was not initialized yet");
-      this.sendResponse(response);
-      return;
-    }
-
-    response.body = response.body || {};
-
-    response.body.scopes =
-      this.pausedScopeChains[args.frameId]?.map((scope) => ({
-        name: scope.type === "closure" ? "CLOSURE" : scope.name || scope.type.toUpperCase(), // for closure type, names are just numbers, so they don't look good, instead we just use name "CLOSURE"
-        variablesReference: this.cdpSession!.adaptCDPObjectId(scope.object.objectId),
-        presentationHint: inferDAPScopePresentationHintFromCDPType(scope.type),
-        expensive: scope.type !== "local", // we only mark local scope as non-expensive as it is the one typically people want to look at and shouldn't have too many objects
-      })) || [];
-    this.sendResponse(response);
-  }
-
-  protected async sourceRequest(
-    response: DebugProtocol.Response & {
-      body: {
-        sourceURL: string;
-        lineNumber1Based: number;
-        columnNumber0Based: number;
-        scriptURL: string;
-      };
-    },
-    args: DebugProtocol.SourceArguments & {
-      fileName: string;
-      line0Based: number;
-      column0Based: number;
-    },
-    request?: DebugProtocol.Request
-  ) {
-    if (!this.cdpSession) {
-      Logger.warn("[DebugAdapter] [sourceRequest] The CDPSession was not initialized yet");
-      this.sendResponse(response);
-      return;
-    }
-
-    response.body = {
-      ...this.cdpSession.findOriginalPosition(args.fileName, args.line0Based, args.column0Based),
-    };
-    this.sendResponse(response);
-  }
-
-  protected async variablesRequest(
-    response: DebugProtocol.Response,
-    args: DebugProtocol.VariablesArguments
-  ): Promise<void> {
-    if (!this.cdpSession) {
-      Logger.warn("[DebugAdapter] [variablesRequest] The CDPSession was not initialized yet");
-      this.sendResponse(response);
-      return;
-    }
-
-    response.body = response.body || {};
-    response.body.variables = [];
-
-    try {
-      if (args.filter !== "indexed" && args.filter !== "named") {
-        response.body.variables = await this.cdpSession.getVariable(args.variablesReference);
-      } else if (args.filter === "indexed") {
-        const stringified = "" + getArraySlots;
-
-        const partialValue = await this.cdpSession!.sendCDPMessage("Runtime.callFunctionOn", {
-          functionDeclaration: stringified,
-          objectId: this.cdpSession.convertDAPObjectIdToCDP(args.variablesReference),
-          arguments: [args.start, args.count].map((value) => ({ value })),
-        });
-
-        const properties = await this.cdpSession.getVariable(
-          this.cdpSession.adaptCDPObjectId(partialValue.result.objectId)
-        );
-
-        response.body.variables = properties;
-      } else if (args.filter === "named") {
-        // We do nothing for named variables. We set 'named' and 'indexed' only for arrays in variableStore
-        // so the 'named' here means "display chunks" (which is handled by Debugger). If we'd get the properties
-        // here we would get all indexed properties even when passing `nonIndexedPropertiesOnly: true` param
-        // to Runtime.getProperties. I assume that this property just does not work yet as it's marked as experimental.
-      }
-    } catch (e) {
-      Logger.error("[CDP] Failed to retrieve variable", e);
-      response.success = false;
-      response.body.error = {
-        id: ERROR_RESPONSE_FAIL_TO_RETRIEVE_VARIABLE_ID,
-        format: "Failed to retrieve variable",
-      };
-    }
-
-    this.sendResponse(response);
-  }
-
-  protected async continueRequest(
-    response: DebugProtocol.ContinueResponse,
-    args: DebugProtocol.ContinueArguments
-  ): Promise<void> {
-    if (!this.cdpSession) {
-      Logger.warn("[DebugAdapter] [continueRequest] The CDPSession was not initialized yet");
-      this.sendResponse(response);
-      return;
-    }
-
-    await this.cdpSession.sendCDPMessage("Debugger.resume", { terminateOnResume: false });
-    this.sendResponse(response);
-    this.sendEvent(new Event("RNIDE_continued"));
-  }
-
-  protected async nextRequest(
-    response: DebugProtocol.NextResponse,
-    args: DebugProtocol.NextArguments
-  ): Promise<void> {
-    if (!this.cdpSession) {
-      Logger.warn("[DebugAdapter] [nextRequest] The CDPSession was not initialized yet");
-      this.sendResponse(response);
-      return;
-    }
-
-    await this.cdpSession.sendCDPMessage("Debugger.stepOver", {});
-    this.sendResponse(response);
-  }
-
-  protected disconnectRequest(
-    response: DebugProtocol.DisconnectResponse,
-    args: DebugProtocol.DisconnectArguments
-  ): void {
-    if (!this.cdpSession) {
-      Logger.warn("[DebugAdapter] [disconnectRequest] The CDPSession was not initialized yet");
-      this.sendResponse(response);
-      return;
-    }
-
-    this.cdpSession.closeConnection();
-    this.sendResponse(response);
-  }
-
-  protected async evaluateRequest(
-    response: DebugProtocol.EvaluateResponse,
-    args: DebugProtocol.EvaluateArguments
-  ): Promise<void> {
-    if (!this.cdpSession) {
-      Logger.warn("[DebugAdapter] [evaluateRequest] The CDPSession was not initialized yet");
-      this.sendResponse(response);
-      return;
-    }
-
-    const cdpResponse = await this.cdpSession!.sendCDPMessage("Runtime.evaluate", {
-      expression: args.expression,
-    });
-    const remoteObject = cdpResponse.result;
-    const stringValue = inferDAPVariableValueForCDPRemoteObject(remoteObject);
-
-    response.body = response.body || {};
-    response.body.result = stringValue;
-    response.body.variablesReference = 0;
-    if (remoteObject.type === "object") {
-      const dapID = this.cdpSession.adaptCDPObjectId(remoteObject.objectId);
-      response.body.type = "object";
-      response.body.variablesReference = dapID;
-    }
-    this.sendResponse(response);
-  }
-
   private async ping() {
-    if (!this.cdpSession) {
-      Logger.warn("[DebugAdapter] [ping] The CDPSession was not initialized yet");
-      return;
-    }
-
-    try {
-      const res = await this.cdpSession.sendCDPMessage("Runtime.evaluate", {
-        expression: "('ping')",
-      });
-
-      const { result } = res;
-
-      if (result.value === "ping") {
-        this.sendEvent(new Event("RNIDE_pong"));
-      }
-    } catch (_) {
-      /** debugSession is waiting for an event, if it won't get any it will fail after timeout, so we don't need to do anything here */
-    }
+    // if (!this.cdpSession) {
+    //   Logger.warn("[DebugAdapter] [ping] The CDPSession was not initialized yet");
+    //   return;
+    // }
+    // try {
+    //   const res = await this.cdpSession.sendCDPMessage("Runtime.evaluate", {
+    //     expression: "('ping')",
+    //   });
+    //   const { result } = res;
+    //   if (result.value === "ping") {
+    //     this.sendEvent(new Event("RNIDE_pong"));
+    //   }
+    // } catch (_) {
+    //   /** debugSession is waiting for an event, if it won't get any it will fail after timeout, so we don't need to do anything here */
+    // }
   }
 
   protected async customRequest(
@@ -401,25 +97,20 @@ export class DebugAdapter extends DebugSession implements CDPSessionDelegate {
   ) {
     switch (command) {
       case "RNIDE_connect_cdp_debugger":
+        if (this.cdpDebugSession) {
+          debug.stopDebugging(this.cdpDebugSession);
+          this.cdpDebugSession = null;
+        }
         this.connectJSDebugger(args);
         break;
       case "RNIDE_log_message":
         this.logCustomMessage(args.message, args.type, args.source);
         break;
       case "RNIDE_startProfiling":
-        if (this.cdpSession) {
-          await this.cdpSession.startProfiling();
-          this.sendEvent(new Event("RNIDE_profilingCPUStarted"));
-        }
+        await this.cdpDebugSession?.customRequest("RNIDE_startProfiling", {});
         break;
       case "RNIDE_stopProfiling":
-        if (this.cdpSession) {
-          const profile = await this.cdpSession.stopProfiling();
-          const fileName = `profile-${Date.now()}.cpuprofile`;
-          const filePath = path.join(os.tmpdir(), fileName);
-          await fs.promises.writeFile(filePath, JSON.stringify(profile));
-          this.sendEvent(new Event("RNIDE_profilingCPUStopped", { filePath }));
-        }
+        await this.cdpDebugSession?.customRequest("RNIDE_stopProfiling", {});
         break;
       case "RNIDE_ping":
         this.ping();
@@ -429,6 +120,4 @@ export class DebugAdapter extends DebugSession implements CDPSessionDelegate {
     }
     this.sendResponse(response);
   }
-
-  //#endregion
 }
